Split reframe tests into multiple files
